separate modules for compiler and plugin
this change makes it so that I can separately test the classpath
manipulation games to reproduce and the attempt to initialize
from Scala. it turns out the problem is reproducible in pure scala!
which is great because it probably means I don't have to think about
dynamic classloading ever again.